+++ makerpm.pl
@@ -172,8 +172,9 @@ fi
 inetd=`grep "^inetd=" /etc/webmin/miniserv.conf 2>/dev/null | sed -e 's/inetd=//g'`
 startafter=0
 
-# Fix old versions of Webmin that might kill the UI process on upgrade
-if [ -d /etc/webmin ]; then
+# Fix old versions of Webmin that might kill the UI
+# process on upgrade, unless it's already fixed version
+if [ -d /etc/webmin ] && [ ! -f "/etc/webmin/stop-init" ]; then
 	cat >/etc/webmin/stop 2>/dev/null <<'EOD'
 #!/bin/sh
 pidfile=`grep "^pidfile=" /etc/webmin/miniserv.conf | sed -e 's/pidfile=//g'`
@@ -188,7 +189,7 @@ EOD
 fi
 
 if [ "\$1" != 1 ]; then
-	# Upgrading the RPM, so stop the old webmin properly
+	# Upgrading the RPM, so stop the old Webmin properly
 	if [ "\$inetd" != "1" ]; then
 		kill -0 `cat /var/webmin/miniserv.pid 2>/dev/null` 2>/dev/null
 		if [ "\$?" = 0 ]; then
